+++ src/main/java/io/github/jamalam360/honk/compatibility/waila/WailaCompatibility.java
@@ -25,6 +25,7 @@
 package io.github.jamalam360.honk.compatibility.waila;
 
 import io.github.jamalam360.honk.HonkInit;
+import io.github.jamalam360.honk.block.feeder.FeederBlock;
 import io.github.jamalam360.honk.entity.egg.EggEntity;
 import io.github.jamalam360.honk.entity.honk.HonkEntity;
 import mcp.mobius.waila.api.IRegistrar;
@@ -45,5 +46,7 @@ public class WailaCompatibility implements IWailaPlugin {
 		registrar.addConfig(ONLY_SHOW_GENES_ON_SNEAK, false);
 		registrar.addComponent(HonkOverride.INSTANCE, TooltipPosition.BODY, HonkEntity.class);
 		registrar.addComponent(HonkOverride.INSTANCE, TooltipPosition.BODY, EggEntity.class);
+		registrar.addComponent(FeederOverride.INSTANCE, TooltipPosition.BODY, FeederBlock.class);
+		registrar.addBlockData(FeederOverride.INSTANCE, FeederBlock.class);
 	}
 }