+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 - Add an adventure mode check for the feeder.
 - Fix a couple of issues with the feeder.
+- Add creative feeder.
+- Add WAILA supoprt for the feeder.
 
 [Full Changelog](https://github.com/JamCoreModding/Honk/compare/0.1.0-beta.10...0.1.0-beta.11)